+++ vehicle/src/Vehicle/Data/Builtin/Loss/Type.hs
@@ -5,6 +5,9 @@ module Vehicle.Data.Builtin.Loss.Type
   )
 where
 
+import Control.Monad.Writer.Strict (Writer, runWriter, tell)
+import Data.List.NonEmpty (NonEmpty (..))
+import Data.Monoid (Any (..))
 import Data.Proxy (Proxy (..))
 import Vehicle.Backend.ITP.Core (ComparisonType (..), decideIfPointwiseOrReductionComparison)
 import Vehicle.Compile.Prelude
@@ -23,7 +26,7 @@ import Vehicle.Data.Builtin.Standard
     DerivedFunction (..),
   )
 import Vehicle.Data.Code.DSL
-import Vehicle.Data.Code.Interface.Args (IsArgs (..), StackTensorArgs (..))
+import Vehicle.Data.Code.Interface.Args (IsArgs (..), StackTensorArgs (..), VectorLitArgs (..))
 import Vehicle.Data.DSL
 import Vehicle.Data.DifferentiableLogic (TensorDifferentiableLogicField (..))
 import Vehicle.Data.Variable.Free.Context (MonadFreeContext (..))
@@ -105,7 +108,7 @@ typeStandardConstructor c = case c of
   UnitLiteral -> typeOfBuiltinConstructor c
   IndexLiteral {} -> typeOfBuiltinConstructor c
   NatLiteral {} -> typeOfBuiltinConstructor c
-  -- TODO: known bug. Have to separate out veclit into a specialised rat type that takes the max over gradients
+  -- Only reached when no `Real` is written in the element type; see `VectorLiteralWithGradients`.
   VectorLiteral -> typeOfBuiltinConstructor c
   BoolTensorLiteral {} -> typeOfBuiltinConstructor c
   NatTensorLiteral {} -> typeOfBuiltinConstructor c
@@ -114,6 +117,7 @@ typeLossBuiltinConstructor :: LossBuiltinConstructor -> DSLExpr (LossBuiltin mod
 typeLossBuiltinConstructor = \case
   WithGradients -> tGradient
   WithoutGradients -> tGradient
+  VectorLiteralWithGradients -> typeOfVectorLiteralWithGradients
 
 typeLossTypeClass :: LossBuiltinTypeClass -> DSLExpr (LossBuiltin mode)
 typeLossTypeClass = \case
@@ -249,6 +253,17 @@ typeOfStackTensorWithGradients =
     finalTensor d ds =
       explLam "g" tGradient $ \g ->
         tTensor (tRat .@@ [g]) (dimCons d ds)
+
+typeOfVectorLiteralWithGradients :: DSLExpr (LossBuiltin mode)
+typeOfVectorLiteralWithGradients =
+  forAll "t" (tGradient .~> type0) $ \t ->
+    forAllDim Relevant $ \d ->
+      iterate (tGradient ~> type0) (accumulateGradient (\g -> t .@@ [g])) d (finalVector t d)
+        @@ [withoutGradients]
+  where
+    finalVector t d =
+      explLam "g" tGradient $ \g ->
+        tVector (t .@@ [g]) d
 
 accumulateGradient ::
   (DSLExpr (LossBuiltin mode) -> DSLExpr (LossBuiltin mode)) ->
@@ -456,6 +471,14 @@ convertToLossBuiltins decl = do
             QuantifyRecord {} -> developerError "quantifiers should have been eliminated"
         BuiltinConstructor c -> return $ case c of
           BoolTensorLiteral {} -> castWith FromBoolTensorTC (sameConstructor c)
+          -- Typed separately when the element type has a `Real` to carry a gradient.
+          VectorLiteral -> case getExpr accessSpine args of
+            Just vecArgs -> case gradientFamily p (vecLitType vecArgs) of
+              Just family ->
+                normAppList (Builtin p (LossBuiltinConstructor VectorLiteralWithGradients)) $
+                  mkExpr accessSpine vecArgs {vecLitType = family}
+              Nothing -> sameConstructor c
+            Nothing -> developerError "Malformed type-checked vector literal"
           _ -> sameConstructor c
         BuiltinType t -> case t of
           RatType -> return $ normAppList (Builtin p $ StandardBuiltinType t) [explicitIrrelevant (mkRatTypeArg p)]
@@ -477,6 +500,33 @@ convertToLossBuiltins decl = do
         prependHoles n xs = replicate n (implicit $ Hole p "_") <> xs
         castWith f original = normAppList (Builtin p $ LossBuiltinTypeClassOp f) [explicit original]
         convertTo n f = return $ normAppList (Builtin p f) (prependHoles n args)
+
+-- | The family `\g -> t[g]` giving every `Real` in `t` the gradient `g`, if `t` has one. The binder
+-- is irrelevant, so removing irrelevant code after typing turns the family back into `t`.
+gradientFamily :: Provenance -> Expr (LossBuiltin mode) -> Maybe (Expr (LossBuiltin mode))
+gradientFamily p t = do
+  let (body, Any hasReal) = runWriter $ go 0 (liftDBIndices 1 t)
+  if hasReal then Just (Lam p binder body) else Nothing
+  where
+    binder = Binder (BinderDisplayForm (OnlyName "g" p) True) Explicit Irrelevant (Builtin p (LossBuiltinType GradientType))
+
+    go :: Int -> Expr (LossBuiltin mode) -> Writer Any (Expr (LossBuiltin mode))
+    go depth expr = case expr of
+      App fun@(Builtin _ (StandardBuiltinType RatType)) (gradient :| []) -> do
+        tell (Any True)
+        return $ App fun [BoundVar p (Ix depth) <$ gradient]
+      App fun args -> App <$> go depth fun <*> traverse (traverse (go depth)) args
+      Pi p' b res -> Pi p' <$> traverse (go depth) b <*> go (depth + 1) res
+      Lam p' b body -> Lam p' <$> traverse (go depth) b <*> go (depth + 1) body
+      Let p' bound b body -> Let p' <$> go depth bound <*> traverse (go depth) b <*> go (depth + 1) body
+      Record p' ident fields -> Record p' ident <$> traverseRecordFields (go depth) fields
+      RecordProj p' recordType value field -> RecordProj p' <$> go depth recordType <*> go depth value <*> pure field
+      Universe {} -> return expr
+      FreeVar {} -> return expr
+      BoundVar {} -> return expr
+      Hole {} -> return expr
+      Meta {} -> return expr
+      Builtin {} -> return expr
 
 restrictDecidabilityDeclType ::
   forall m mode.